% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csplot_common_support.R
\name{csplot_common_support}
\alias{csplot_common_support}
\title{Internal helper to plot common support}
\usage{
csplot_common_support(ps, D)
}
\arguments{
\item{ps}{Numeric propensity score vector}

\item{D}{Binary treatment indicator}
}
\value{
A recorded base plot object
}
\description{
Internal helper to plot common support
}
\keyword{internal}
