% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPWate.R
\name{IPWate}
\alias{IPWate}
\title{Estimate Average Treatment Effect (ATE) using the Inverse Propensity Weighting
method and machine learning to estimate the propensity score}
\usage{
IPWate(
  Y,
  X,
  D,
  ML = c("RF", "Logit_lasso", "CIF", "XGB", "CB", "grf", "SL"),
  pscore = NULL,
  SL.library = c("SL.ranger"),
  polynomial.Logit_lasso = 1,
  polynomial.Lasso = 1,
  polynomial.Ridge = 1,
  polynomial.OLS = 1,
  polynomial.NLLS_exp = 1,
  polynomial.loglin = 1,
  rf.cf.ntree = 500,
  rf.depth = NULL,
  mtry = max(floor(ncol(X)/3), 1),
  cf.depth = Inf,
  xgb.nrounds = 200,
  xgb.max.depth = 6,
  cb.iterations = 500,
  cb.depth = 6,
  torch.epochs = 50,
  torch.hidden_units = c(64, 32),
  torch.lr = 0.01,
  torch.dropout = 0.2,
  intercept = TRUE
)
}
\arguments{
\item{Y}{is a vector containing the outcome}

\item{X}{is a dataframe containing all the covariates}

\item{D}{the treatment indicator}

\item{ML}{is a string specifying which machine learner to use among
Random Forests, CIF, XGB, CB, Logit Lasso, grf or SuperLearner}

\item{pscore}{vector with treatment probabilities if known (e.g. in an RCT).
Otherwise leave NULL so that they are estimated using ML}

\item{SL.library}{string vector with learners to use in SuperLearner}

\item{polynomial.Logit_lasso}{degree of polynomial to be fitted when using
Logit Lasso. 1 just fits the input X. 2 squares all variables and adds all
pairwise interactions. 3 squares and cubes all variables and adds all
pairwise and threewise interactions...}

\item{polynomial.Lasso}{degree of polynomial to be fitted when using Lasso,
see polynomial.Logit_lasso for more info}

\item{polynomial.Ridge}{degree of polynomial to be fitted when using Ridge,
see polynomial.Logit_lasso for more info}

\item{polynomial.OLS}{degree of polynomial to be fitted when using OLS,
see polynomial.Logit_lasso for more info}

\item{polynomial.NLLS_exp}{degree of polynomial to be fitted when using
NLLS_exp, see polynomial.Logit_lasso for more info}

\item{polynomial.loglin}{degree of polynomial to be fitted when using
loglin, see polynomial.Logit_lasso for more info}

\item{rf.cf.ntree}{how many trees should be grown when using RF or CIF}

\item{rf.depth}{how deep should trees be grown in RF (NULL is full depth,
the default from ranger)}

\item{mtry}{number of variables to possibly split at in each node for RF or CIF}

\item{cf.depth}{how deep should trees be grown in CIF (Inf is default from partykit)}

\item{xgb.nrounds}{number of boosting rounds to use in XGB}

\item{xgb.max.depth}{maximum tree depth in XGB}

\item{cb.iterations}{maximum number of trees that can be built in CB}

\item{cb.depth}{depth of trees in CB}

\item{torch.epochs}{number of training epochs for the Torch neural network}

\item{torch.hidden_units}{numeric vector specifying the number of neurons in
each hidden layer of the Torch neural network}

\item{torch.lr}{learning rate for the Torch optimizer}

\item{torch.dropout}{dropout rate for regularization in the Torch neural
network (between 0 and 1)}

\item{intercept}{logical; should an intercept be included? Default TRUE.
Only applies to OLS, Lasso, Ridge, and loglin.}
}
\value{
dataframe with estimates
}
\description{
\code{IPWate} estimates the ATE by estimating the propensity score by Machine Learning
and implementing a plug-in IPW estimator
}
\examples{
n <- 1000
X <- rnorm(n)
D <- runif(n) >= 0.5
Y0 <- X + rnorm(n)
Y1 <- 5 + X + 5*X + rnorm(n)
Y <- Y1*D + Y0*(1-D)
IPWate(Y,X,D, ML = "RF")

}
