% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLSate.R
\name{OLSate}
\alias{OLSate}
\title{Estimate Average Treatment Effect (ATE) in Randomized Control Trials (RCTs) with OLS}
\usage{
OLSate(Y, X, D)
}
\arguments{
\item{Y}{is a vector containing the outcome}

\item{X}{is a dataframe containing all the covariates}

\item{D}{the treatment indicator}
}
\value{
dataframe with estimates and standard errors
}
\description{
\code{OLSate} estimates the ATE with data from RCTs. Reports both the estimate which
is unadjusted by covariates and the estimate adjusted by covariates. The adjustment
is made recentering the covariates and interacting them with the treatment so that
precision is improved even if the model is not correctly specified.
}
\examples{
n <- 1000
X <- rnorm(n)
D <- runif(n) >= 0.5
Y0 <- X + rnorm(n)
Y1 <- 5 + X + rnorm(n)
Y <- Y1*D + Y0*(1-D)
OLSate(Y,X,D)

}
