% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directMLate.R
\name{directMLate}
\alias{directMLate}
\title{Estimate Average Treatment Effect (ATE) using machine learning to
estimate the conditional expectations}
\usage{
directMLate(Y, X, D, ML = c("Lasso", "Ridge", "RF", "XGB"), polynomial = 1)
}
\arguments{
\item{Y}{is a vector containing the outcome}

\item{X}{is a dataframe containing all the covariates}

\item{D}{the treatment indicator}

\item{ML}{is a string specifying which machine learner to use among
Lasso, Ridge, Random Forests or XGBoosting}

\item{polynomial}{degree of polynomial to be fitted when using Lasso, Ridge
or Logit Lasso. 1 just fits the input X. 2 squares all variables and adds
all pairwise interactions. 3 squares and cubes all variables and adds all
pairwise and threewise interactions...}
}
\value{
dataframe with estimates
}
\description{
\code{directMLate} estimates the ATE by estimating the conditional expectation
of Y given X for the treated: mu_1(X) and for the control: mu_0(X) and
estimating the expected difference among them.
}
\examples{
n <- 1000
X <- rnorm(n)
D <- runif(n) >= 0.5
Y0 <- X + rnorm(n)
Y1 <- 5 + X + 5*X + rnorm(n)
Y <- Y1*D + Y0*(1-D)
directMLate(Y,X,D, ML = "RF")

}
