% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPWate.R
\name{IPWate}
\alias{IPWate}
\title{Estimate Average Treatment Effect (ATE) using the Inverse Propensity Weighting
method and machine learning to estimate the propensity score}
\usage{
IPWate(
  Y,
  X,
  D,
  ML = c("RF", "Lasso_logit", "grf", "SL"),
  pscore = NULL,
  polynomial = 1,
  SL.library = c("SL.ranger")
)
}
\arguments{
\item{Y}{is a vector containing the outcome}

\item{X}{is a dataframe containing all the covariates}

\item{D}{the treatment indicator}

\item{ML}{is a string specifying which machine learner to use among
Random Forests or Lasso logit}

\item{pscore}{vector with treatment probabilities if known (e.g. in an RCT).
Otherwise leave NULL so that they are estimated using MLps}

\item{polynomial}{degree of polynomial to be fitted when using Lasso, Ridge
or Logit Lasso. 1 just fits the input X. 2 squares all variables and adds
all pairwise interactions. 3 squares and cubes all variables and adds all
pairwise and threewise interactions...}
}
\value{
dataframe with estimates
}
\description{
\code{IPWate} estimates the ATE by estimating the propensity score by Machine Learning
and implementing a plug-in IPW estimator
}
\examples{
n <- 1000
X <- rnorm(n)
D <- runif(n) >= 0.5
Y0 <- X + rnorm(n)
Y1 <- 5 + X + 5*X + rnorm(n)
Y <- Y1*D + Y0*(1-D)
IPWate(Y,X,D, ML = "RF")

}
