% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRate.R
\name{LRate}
\alias{LRate}
\title{Estimate Average Treatment Effect (ATE) using a locally robust moment
and machine learning methods}
\usage{
LRate(
  Y,
  X,
  D,
  MLreg = c("Lasso", "Ridge", "RF", "XGB"),
  MLps = c("RF", "Logit_lasso"),
  polynomial = 1,
  CF = TRUE,
  K = 2
)
}
\arguments{
\item{Y}{is a vector containing the outcome}

\item{X}{is a dataframe containing all the covariates}

\item{D}{the treatment indicator}

\item{MLreg}{is a string specifying which machine learner to use among
Lasso, Ridge, Random Forests or XGBoosting to estimate response functions}

\item{MLps}{is a string specifying which machine learner to use among
Random Forests or Logit Lasso to estimate propensity scores}

\item{polynomial}{degree of polynomial to be fitted when using Lasso, Ridge
or Logit Lasso. 1 just fits the input X. 2 squares all variables and adds
all pairwise interactions. 3 squares and cubes all variables and adds all
pairwise and threewise interactions...}

\item{CF}{whether cross-fitting should be used}

\item{K}{number of splits to use for cross-fitting}
}
\value{
dataframe with estimates and standard errors (if cross-fitting is employed)
}
\description{
\code{LRate} estimates the ATE by using locally robust/orthogonal scores
which have the property of double robustness and for which inference
is available even when using machine learning if cross-fitting is employed
}
\examples{
n <- 1000
X <- rnorm(n)
D <- runif(n) >= 0.5
Y0 <- X + rnorm(n)
Y1 <- 5 + X + 5*X + rnorm(n)
Y <- Y1*D + Y0*(1-D)
LRate(Y,X,D, MLreg = "RF", MLps = "Logit_lasso")

}
